Trait ssc::ast::visit::VisitMut

pub trait VisitMut<'a>: Sized {

    // Provided methods
    fn enter_node(&mut self, _kind: AstType) { ... }
    fn leave_node(&mut self, _kind: AstType) { ... }
    fn visit_root(&mut self, root: &mut Root<'a>) { ... }
    fn visit_fragment(&mut self, fragment: &mut Fragment<'a>) { ... }
    fn visit_fragment_node(&mut self, node: &mut FragmentNode<'a>) { ... }
    fn visit_text(&mut self, text: &mut Text<'a>) { ... }
    fn visit_tag(&mut self, tag: &mut Tag<'a>) { ... }
    fn visit_expression_tag(&mut self, expression_tag: &mut ExpressionTag<'a>) { ... }
    fn visit_html_tag(&mut self, html_tag: &mut HtmlTag<'a>) { ... }
    fn visit_const_tag(&mut self, const_tag: &mut ConstTag<'a>) { ... }
    fn visit_debug_tag(&mut self, debug_tag: &mut DebugTag<'a>) { ... }
    fn visit_render_tag(&mut self, render_tag: &mut RenderTag<'a>) { ... }
    fn visit_element(&mut self, element: &mut Element<'a>) { ... }
    fn visit_component(&mut self, component: &mut Component<'a>) { ... }
    fn visit_title_element(&mut self, title_element: &mut TitleElement<'a>) { ... }
    fn visit_slot_element(&mut self, slot_element: &mut SlotElement<'a>) { ... }
    fn visit_regular_element(
        &mut self,
        regular_element: &mut RegularElement<'a>
    ) { ... }
    fn visit_svelte_body(&mut self, svelte_body: &mut SvelteBody<'a>) { ... }
    fn visit_svelte_component(
        &mut self,
        svelte_component: &mut SvelteComponent<'a>
    ) { ... }
    fn visit_svelte_document(
        &mut self,
        svelte_document: &mut SvelteDocument<'a>
    ) { ... }
    fn visit_svelte_element(&mut self, svelte_element: &mut SvelteElement<'a>) { ... }
    fn visit_svelte_fragment(
        &mut self,
        svelte_fragment: &mut SvelteFragment<'a>
    ) { ... }
    fn visit_svelte_head(&mut self, svelte_head: &mut SvelteHead<'a>) { ... }
    fn visit_svelte_options_raw(
        &mut self,
        svelte_options_raw: &mut SvelteOptionsRaw<'a>
    ) { ... }
    fn visit_svelte_self(&mut self, svelte_self: &mut SvelteSelf<'a>) { ... }
    fn visit_svelte_window(&mut self, svelte_window: &mut SvelteWindow<'a>) { ... }
    fn visit_block(&mut self, block: &mut Block<'a>) { ... }
    fn visit_each_block(&mut self, each_block: &mut EachBlock<'a>) { ... }
    fn visit_if_block(&mut self, if_block: &mut IfBlock<'a>) { ... }
    fn visit_await_block(&mut self, await_block: &mut AwaitBlock<'a>) { ... }
    fn visit_key_block(&mut self, key_block: &mut KeyBlock<'a>) { ... }
    fn visit_snippet_block(&mut self, snippet_block: &mut SnippetBlock<'a>) { ... }

}

Syntax tree traversal to mutate an exclusive borrow of a syntax tree in place a syntax tree in place.

Provided Methods

fn enter_node(&mut self, _kind: AstType)

fn leave_node(&mut self, _kind: AstType)

fn visit_root(&mut self, root: &mut Root<'a>)

fn visit_fragment(&mut self, fragment: &mut Fragment<'a>)

fn visit_fragment_node(&mut self, node: &mut FragmentNode<'a>)

fn visit_text(&mut self, text: &mut Text<'a>)

fn visit_tag(&mut self, tag: &mut Tag<'a>)

fn visit_expression_tag(&mut self, expression_tag: &mut ExpressionTag<'a>)

fn visit_html_tag(&mut self, html_tag: &mut HtmlTag<'a>)

fn visit_const_tag(&mut self, const_tag: &mut ConstTag<'a>)

fn visit_debug_tag(&mut self, debug_tag: &mut DebugTag<'a>)

fn visit_render_tag(&mut self, render_tag: &mut RenderTag<'a>)

fn visit_element(&mut self, element: &mut Element<'a>)

fn visit_component(&mut self, component: &mut Component<'a>)

fn visit_title_element(&mut self, title_element: &mut TitleElement<'a>)

fn visit_slot_element(&mut self, slot_element: &mut SlotElement<'a>)

fn visit_regular_element(&mut self, regular_element: &mut RegularElement<'a>)

fn visit_svelte_body(&mut self, svelte_body: &mut SvelteBody<'a>)

fn visit_svelte_component(&mut self, svelte_component: &mut SvelteComponent<'a>)

fn visit_svelte_document(&mut self, svelte_document: &mut SvelteDocument<'a>)

fn visit_svelte_element(&mut self, svelte_element: &mut SvelteElement<'a>)

fn visit_svelte_fragment(&mut self, svelte_fragment: &mut SvelteFragment<'a>)

fn visit_svelte_head(&mut self, svelte_head: &mut SvelteHead<'a>)

fn visit_svelte_options_raw( &mut self, svelte_options_raw: &mut SvelteOptionsRaw<'a> )

fn visit_svelte_self(&mut self, svelte_self: &mut SvelteSelf<'a>)

fn visit_svelte_window(&mut self, svelte_window: &mut SvelteWindow<'a>)

fn visit_block(&mut self, block: &mut Block<'a>)

fn visit_each_block(&mut self, each_block: &mut EachBlock<'a>)

fn visit_if_block(&mut self, if_block: &mut IfBlock<'a>)

fn visit_await_block(&mut self, await_block: &mut AwaitBlock<'a>)

fn visit_key_block(&mut self, key_block: &mut KeyBlock<'a>)

fn visit_snippet_block(&mut self, snippet_block: &mut SnippetBlock<'a>)

Object Safety

This trait is not object safe.