Module ssc::diagnostics

Structs

DiagnosticService

GraphicalReportHandler

GraphicalTheme

Theme used by GraphicalReportHandler to render fancy Diagnostic reports.

LabeledSpan

A labeled [SourceSpan].

NamedSource

Utility struct for when you have a regular [SourceCode] type that doesn’t implement name. For example String. Or if you want to override the name returned by the SourceCode.

OxcDiagnostic

OxcDiagnosticInner

Type Aliases

DiagnosticSender

DiagnosticTuple

Error

Result

Severity